Combine GameState into Game
This is a massive refactor that includes migrations which drop all currently ongoing game states. Rerunning this in production after actual games are running will not be possible. All fields (roll, roll_count, pieces, current_player) in GameState are added into Game. GameState is dropped from the database and the FK references are remapped into Game.

In addition, the Game and Player structs are moved from the Networking namespace into Logic. Further refactoring to separate game logic from the storage model is still necessary, but this refactor allows continuing implementation for the missing features, which were previously unwieldy. As an example, checking whether the current player should be allowed to do the raise action requires data from both the old GameState (namely, piece positions and the roll) and Game (namely, whether or not the player has already done the raise action).